Rename storage-init to initoverlayfs-setup
storage-init no longer is an init process, nor does it initialize
storage, so best off renaming.
