+++ lib/dracut/modules.d/81initoverlayfs/module-setup.sh
@@ -10,7 +10,7 @@ depends() {
 }
 
 install() {
-    inst_multiple -o /etc/initoverlayfs.conf /usr/sbin/storage-init \
+    inst_multiple -o /etc/initoverlayfs.conf /usr/sbin/initoverlayfs-setup \
       "$systemdsystemunitdir/pre-initoverlayfs.target" \
       "$systemdsystemunitdir/pre-initoverlayfs.service" \
       "$systemdsystemunitdir/pre-initoverlayfs-switch-root.service"
